+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - skip exists jm title
 - fix exists jm title bug
 - catch method download_album exception
+- update jm option config
+- enable jm loguru log record
 
 ## v1.1.8-beta.1.240624
 ### 2024/6/24